+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Hide condition fields in replication settings when `each_n` and `each_s` are null or undefined, [PR-106](https://github.com/reductstore/web-console/pull/106)
-- Use streams for uploading files, [PR-113](https://github.com/reductstore/web-console/pull/113)
+- Use streams when downloading large records, [PR-113](https://github.com/reductstore/web-console/pull/113)
 
 ## [1.10.1] - 2025-05-20